+++ config/filament-user.php
@@ -1,0 +1,6 @@
+<?php
+
+// config for Kgalanos/FilamentUser
+return[
+    'IsUsernameEditable' => true,
+];