+++ src/Filament/Pages/Auth/Login.php
@@ -1,0 +1,95 @@
+<?php
+
+namespace Kgalanos\FilamentUser\Filament\Pages\Auth;
+
+use DanHarrin\LivewireRateLimiting\Exceptions\TooManyRequestsException;
+use Filament\Facades\Filament;
+use Filament\Forms\Components\Component;
+use Filament\Forms\Components\TextInput;
+use Filament\Forms\Form;
+use Filament\Http\Responses\Auth\Contracts\LoginResponse;
+use Filament\Notifications\Notification;
+use Illuminate\Validation\ValidationException;
+
+class Login extends \Filament\Pages\Auth\Login
+{
+    public function form(Form $form): Form
+    {
+        return $form
+            ->schema([
+                $this->getUsernameOrEmailFormComponent(),
+                $this->getPasswordFormComponent(),
+                $this->getRememberFormComponent(),
+            ])
+            ->statePath('data');
+    }
+
+    protected function getUsernameOrEmailFormComponent(): Component
+    {
+        return TextInput::make('UsernameOrEmail')
+            ->label(__('Username or email'))
+            ->required()
+            ->autocomplete()
+            ->autofocus();
+    }
+
+    public function authenticate(): ?LoginResponse
+    {
+        try {
+            $this->rateLimit(5);
+        } catch (TooManyRequestsException $exception) {
+            Notification::make()
+                ->title(__('filament-panels::pages/auth/login.notifications.throttled.title', [
+                    'seconds' => $exception->secondsUntilAvailable,
+                    'minutes' => ceil($exception->secondsUntilAvailable / 60),
+                ]))
+                ->body(array_key_exists('body', __('filament-panels::pages/auth/login.notifications.throttled') ?: []) ? __('filament-panels::pages/auth/login.notifications.throttled.body', [
+                    'seconds' => $exception->secondsUntilAvailable,
+                    'minutes' => ceil($exception->secondsUntilAvailable / 60),
+                ]) : null)
+                ->danger()
+                ->send();
+
+            return null;
+        }
+
+        $data = $this->form->getState();
+
+        if (
+            ! Filament::auth()->attempt($this->getCredentialsUsernameFromFormData($data), $data['remember'] ?? false) &&
+            ! Filament::auth()->attempt($this->getCredentialsEmailFromFormData($data), $data['remember'] ?? false)
+        ) {
+            throw ValidationException::withMessages([
+                'data.UsernameOrEmail' => __('filament-panels::pages/auth/login.messages.failed'),
+            ]);
+        }
+
+        session()->regenerate();
+
+        return app(LoginResponse::class);
+    }
+
+    /**
+     * @param  array<string, mixed>  $data
+     * @return array<string, mixed>
+     */
+    protected function getCredentialsEmailFromFormData(array $data): array
+    {
+        return [
+            'email' => $data['UsernameOrEmail'],
+            'password' => $data['password'],
+        ];
+    }
+
+    /**
+     * @param  array<string, mixed>  $data
+     * @return array<string, mixed>
+     */
+    protected function getCredentialsUsernameFromFormData(array $data): array
+    {
+        return [
+            'username' => $data['UsernameOrEmail'],
+            'password' => $data['password'],
+        ];
+    }
+}