+++ src/Filament/Widgets/ApplicationInfoWidget.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Kgalanos\FilamentUser\Filament\Widgets;
+
+use Filament\Widgets\Widget;
+
+class ApplicationInfoWidget extends Widget
+{
+    protected static ?int $sort = 0;
+
+    protected static string $view = 'vendor.kgalanos.filament-user.filament.widgets.application-info-widget'; //'application-info-widget';
+}