+++ src/FilamentUser.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Kgalanos\FilamentUser;
+
+class FilamentUser
+{
+}