+++ src/FilamentUserServiceProvider.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace VendorName\Skeleton;
+namespace Kgalanos\FilamentUser;
 
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
-use VendorName\Skeleton\Commands\SkeletonCommand;
+use Kgalanos\FilamentUser\Commands\FilamentUserCommand;
 
-class SkeletonServiceProvider extends PackageServiceProvider
+class FilamentUserServiceProvider extends PackageServiceProvider
 {
     public function configurePackage(Package $package): void
     {
@@ -16,10 +16,9 @@ class SkeletonServiceProvider extends PackageServiceProvider
          * More info: https://github.com/spatie/laravel-package-tools
          */
         $package
-            ->name('skeleton')
-            ->hasConfigFile()
-            ->hasViews()
-            ->hasMigration('create_skeleton_table')
-            ->hasCommand(SkeletonCommand::class);
+            ->name('filament-user')
+            ->hasConfigFile('filament-user')
+            ->hasViews('kgalanos\\filament-user')
+            ->hasMigration('migrate_users_table');
     }
 }