+++ src/Models/User.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Kgalanos\FilamentUser\Models;
+
+// use Illuminate\Contracts\Auth\MustVerifyEmail;
+use Filament\Models\Contracts\FilamentUser;
+use Filament\Models\Contracts\HasAvatar;
+use Filament\Panel;
+use Illuminate\Database\Eloquent\Concerns\HasUlids;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Foundation\Auth\User as Authenticatable;
+use Illuminate\Notifications\Notifiable;
+use Laravel\Sanctum\HasApiTokens;
+
+/**
+ * @property null|string $username
+ * @property null|string $avatar_url
+ */
+class User extends Authenticatable implements FilamentUser, HasAvatar
+{
+    use HasApiTokens, HasFactory, Notifiable, HasUlids;
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array<int, string>
+     */
+    protected $fillable = [
+        'name',
+        'email',
+        'password',
+        'username',
+        'avatar_url',
+        'phone',
+    ];
+
+    /**
+     * The attributes that should be hidden for serialization.
+     *
+     * @var array<int, string>
+     */
+    protected $hidden = [
+        'password',
+        'remember_token',
+    ];
+
+    /**
+     * The attributes that should be cast.
+     *
+     * @var array<string, string>
+     */
+    protected $casts = [
+        'email_verified_at' => 'datetime',
+        'password' => 'hashed',
+    ];
+
+    public function canAccessPanel(Panel $panel): bool
+    {
+        // TODO: Implement canAccessPanel() method.
+        return true;
+    }
+
+    public function getFilamentAvatarUrl(): ?string
+    {
+        // TODO: Implement getFilamentAvatarUrl() method.
+        return $this->avatar_url;
+    }
+
+    public function uniqueIds()
+    {
+        return ['ulid'];
+    }
+}