+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `filament-user` will be documented in this file.
+
+# "Initial Release Version 1.0.0"
+
+last commit in 2023-08-19
+
+* Add fields ulid, username, avatar_url, phone, phone_verified_at
+* You can login with username or email